@@ -1,4 +1,4 @@
-class MpInfo < ActiveRecord::Base
+class MpInfo < ApplicationRecord
   belongs_to :mp, primary_key: :deputy_id, foreign_key: :deputy_id
   def rebellions_fraction
     rebellions.to_f / votes_attended if votes_attended > 0
