@@ -1,4 +1,4 @@
-class Vote < ActiveRecord::Base
+class Vote < ApplicationRecord
   belongs_to :division
   belongs_to :mp, primary_key: :deputy_id, foreign_key: :deputy_id
   def self.find_mp(id)
