@@ -1,3 +1,3 @@
-class Whip < ActiveRecord::Base
+class Whip < ApplicationRecord
   belongs_to :division
 end
